4.21*: bloglist expansion, step9

+++ part4/blog-list/controllers/blogs.js
@@ -63,8 +63,18 @@ blogsRouter.post('/', async (req, res) => {
 })
 
 blogsRouter.delete('/:id', async (req, res) => {
-	await Blog.findByIdAndRemove(req.params.id)
-	res.status(204).end()
+	const decodedToken = jwt.verify(req.token, process.env.SECRET)
+	if (!req.token || !decodedToken.id) {
+		return res.status(401).json({ error: 'token missing or invalid' })
+	}
+	const userid = decodedToken.id
+	const blog = await Blog.findById(req.params.id)
+	if(blog.user.toString() === userid) {
+		await Blog.findByIdAndRemove(req.params.id)
+		res.status(204).end()
+	} else {
+		return res.status(401).json({ error: 'you do not have edit right' })
+	}
 })
 
 blogsRouter.put('/:id', async (req, res) => {
